In build.sh before remove vcpkg_* directories, generate a new archive from vcpkg_installed directory into ~/vcpkg_downloads-<BASE_BRANCH>.zip
Upload all vcpkg_downloads-*.zip with OBT results.

+++ build.sh
@@ -584,8 +584,20 @@ then
         WriteLog "Clean-up /opt/HPCCSystems/lib and plugins directories" "${OBT_BUILD_LOG_FILE}"
         sudo rm -rf /opt/HPCCSystems/lib/ /opt/HPCCSystems/plugins/
         
-        WriteLog "Clean-up build/vcpkg_* directories to save disk space." "${OBT_BUILD_LOG_FILE}"
-        rm -rf ${BUILD_HOME}/vcpkg_downloads ${BUILD_HOME}/vcpkg_buildtrees ${BUILD_HOME}/vcpkg_packages
+        if [[ -d ${BUILD_HOME}/vcpkg_installed  ]]
+        then
+            BASE_VERSION=${BRANCH_ID#candidate-}
+            BASE_VERSION=${BASE_VERSION%.*}
+            [[ "$BASE_VERSION" != "master" ]] && BASE_VERSION=$BASE_VERSION.x
+            WriteLog "Create fresh vcpkg_download-${BASE_VERSION}.zip file" "${OBT_BUILD_LOG_FILE}"
+            mv -fv ~/vcpkg_downloads-${BASE_VERSION}.zip ~/vcpkg_downloads-${BASE_VERSION}-old.zip
+            zip -r ~/vcpkg_downloads-${BASE_VERSION}.zip ${BUILD_HOME}/vcpkg_installed/*
+
+            WriteLog "Clean-up build/vcpkg_* directories to save disk space." "${OBT_BUILD_LOG_FILE}"
+            rm -rf ${BUILD_HOME}/vcpkg_downloads ${BUILD_HOME}/vcpkg_buildtrees ${BUILD_HOME}/vcpkg_packages
+        else
+            WriteLog "VCPKG not used for build." "${OBT_BUILD_LOG_FILE}"
+        fi
 
         res=$( ${SUDO} ${PKG_INST_CMD} ${BUILD_HOME}/$hpcc_package 2>&1 )
         WriteLog "Install package" "${OBT_BUILD_LOG_FILE}"
